% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions1.r
\name{checks_total}
\alias{checks_total}
\title{Total that have completed the course and total employees for reporting purposes}
\usage{
checks_total(staff, course, date)
}
\arguments{
\item{staff}{path to staff excel spreadsheet}

\item{course}{path to course logs downloaded}

\item{date}{cut off date as string in the format "YYYY-MM-DD"}
}
\value{
two values; the first is the number of employees that have completed the course and the second is the total employees
}
\description{
Total that have completed the course and total employees for reporting purposes
}
\examples{
checks total (staff = staff, course = course, date=date)

}
