% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{course_foo}
\alias{course_foo}
\title{Title}
\usage{
course_foo(x, t)
}
\arguments{
\item{x}{course dataset}

\item{t}{cut off date in the form of YYYY-MM-DD, e.g.,2022-07-31}
}
\value{
dataframe with relevant variables
}
\description{
Title
}
