% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions1.r
\name{divisions_completions}
\alias{divisions_completions}
\title{Completion rate for each area (of interest)}
\usage{
divisions_completions(staff, course1, course2, date)
}
\arguments{
\item{staff}{path to staff excel spreadsheet}

\item{course1}{path to course logs downloaded for course1}

\item{course2}{path to course logs downloaded for course2}

\item{date}{cut off date as string in the format "YYYY-MM-DD"}
}
\value{
a dataframe with information on completion per division
}
\description{
Completion rate for each area (of interest)
}
