% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions1.r
\name{individual_completions}
\alias{individual_completions}
\title{Completion rate for each area}
\usage{
individual_completions(staff, course1, course2, date)
}
\arguments{
\item{staff}{path to staff excel spreadsheet}

\item{course1}{path to course logs downloaded for course1}

\item{course2}{path to course logs downloaded for course2}

\item{date}{cut off date as string in the format "YYYY-MM-DD"}
}
\value{
a dataframe with information on completion
}
\description{
Completion rate for each area
}
\examples{
individual_completions(staff = staff, course1 = course1, course2 = course2, date = date)
}
