% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{merge_foo}
\alias{merge_foo}
\title{Merged datasets}
\usage{
merge_foo(x, y)
}
\arguments{
\item{x}{course formatted dataset}

\item{y}{staff formatted dataset}
}
\value{
combined dataset
}
\description{
Merged datasets
}
