% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions1.r
\name{ons_completion_rate}
\alias{ons_completion_rate}
\title{Completion rates for ONS}
\usage{
ons_completion_rate(staff, course, date)
}
\arguments{
\item{staff}{path to staff excel spreadsheet}

\item{course}{path to course logs downloaded}

\item{date}{cut off date as string in the format "YYYY-MM-DD"}
}
\value{
percentage completion rate
}
\description{
Completion rates for ONS
}
\examples{
ons_completion_rate(staff = staff, course = course, date = date)
}
