% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{staff_df}
\alias{staff_df}
\title{Formats the staff spreadsheet to be merged with the e-learning}
\usage{
staff_df(staff)
}
\arguments{
\item{staff}{the excel spreasheet downloaded with the staff counts}
}
\value{
dataframe with relevant variables
}
\description{
Formats the staff spreadsheet to be merged with the e-learning
}
